dialog 분리
