+++ src/App.tsx
@@ -6,7 +6,6 @@ import { BrowserRouter as Router } from 'react-router-dom';
 
 import PostsManagerPage from './pages/PostsManagerPage.tsx';
 
-// QueryClient 인스턴스 생성
 const queryClient = new QueryClient();
 
 const App = () => {
